refactor(computer): move the checkpoint, pause-release and resume bodies (R3, PR-F1d)
Three more sub-task bodies join the machine that names them:
`checkpoint_impl` and its capture (`Effect::SpawnCheckpoint`), the pause
release that keeps the disk (`SpawnRelease { KeepDisk }`), and the in-place
restore of a paused computer with its unwind (`SpawnResume`). Their callers
in `sandbox::{checkpoint,pause,templates,warm}` still drive them, so nothing
changes about what runs or in what order.

The checkpoint half is a pure move — it was already free-standing so all
three of its callers could share one capture path. The pause and resume
halves were `SandboxManager` methods, so they take what they read off the
manager as parameters, exactly as the restore did in PR-F1c;
`preserved_cow_file` becomes a free function of the config and the id, which
is all it ever was.

Carries PR-E's gap 2: a resume failure is two-valued and now the vocabulary
is too. `resume_sandbox` parks at `Paused` when the restore unwound and at
`Failed` when it did not, and the difference is crash safety rather than
cosmetics — `Paused` is a promise that the retained state is whole, which
the restart sweep reads as "reinstate this as resumable" and which lets it
drop the journal naming what a half-allocated computer still holds.
`Event::Stranded` is that second value, shaped like `Frozen`: a failure
class the root superstate fails, which `resuming` deliberately does not
claim. `TaskFailure::stranded` is how a sub-task reports it.
